refactor(service): factorize params service tests
